updated code
updted code with user condoler

+++ src/main/java/com/coderscampus/flightTrack/web/UserController.java
@@ -98,8 +98,15 @@ public class UserController {
 
 	@PostMapping("/users/{userId}/delete")
 	public String postDeleteUser(@PathVariable Integer userId) {
-		refreshTokenService.deleteByUserId(userId);
-		userService.delete(userId);
+
+		Long userCount = userService.getUserCount();
+
+		if (userCount > 1) {
+
+			refreshTokenService.deleteByUserId(userId);
+			userService.delete(userId);
+		}
+
 		return "redirect:/api/v1/users/users";
 	}
